fix copying home board to work as expected
When you had a home board set, if you chose to copy it, it would ask
you this confusing question about the board being in your current
set. Basically it was the same question it would ask if you were
copying a board somewhere downstream in your set. This question has
been replaced by an easier one for the user.